+++ lib/views/chat_view/poll_view.dart
@@ -4,7 +4,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gocast_mobile/views/chat_view/poll_view_state.dart';
 
 class PollView extends ConsumerStatefulWidget {
-  final Int64? streamID;
+  final int? streamID;
 
   const PollView({
     super.key,
