+++ lib/views/course_view/components/course_section.dart
@@ -144,7 +144,7 @@ class CourseSection extends StatelessWidget {
             ref: ref,
             title: course.name,
             tumID: course.tUMOnlineIdentifier,
-            lastLectureId: Int64(course.lastRecordingID),
+            lastLectureId: course.lastRecordingID,
             path: imagePath,
             live: streams.any((stream) => stream.courseID == course.id),
             semester:
